runtime: align generation playback runtime file names
