+++ Sources/SpeakSwiftly/API/Playback.swift
@@ -76,7 +76,7 @@ public extension SpeakSwiftly.Runtime {
 
         await failQueuedRequests(with: cancellationError)
         await failWaitingPlaybackRequests(with: cancellationError)
-        let cancelledPlaybackJobs = await playbackController.shutdown()
+        let cancelledPlaybackJobs = await playbackQueue.shutdown()
         for playbackState in cancelledPlaybackJobs {
             playbackState.execution.continuation.finish(throwing: cancellationError)
             await completePlaybackJob(playbackState.request, result: .failure(cancellationError))
