+++ Sources/SpeakSwiftly/Generation/CloneTranscriptionClient.swift
