+++ Sources/SpeakSwiftly/Generation/ResidentSpeechInputs.swift
