+++ Sources/SpeakSwiftly/Generation/SpeechModelClient.swift
