+++ Sources/SpeakSwiftly/Playback/PlaybackQueue+GenerationAdmission.swift
@@ -1,15 +1,4 @@
-import Foundation
-
-struct PlaybackHooks {
-    let handleEvent: @Sendable (PlaybackEvent, LiveSpeechRequestState) async -> Void
-    let handleEnvironmentEvent: @Sendable (PlaybackEnvironmentEvent, ActiveWorkerRequestSummary?) async -> Void
-    let logEngineReady: @Sendable (LiveSpeechRequestState, Double) async -> Void
-    let logFinished: @Sendable (LiveSpeechRequestState, PlaybackSummary, Double) async -> Void
-    let completeJob: @Sendable (LiveSpeechRequestState, Result<SpeakSwiftly.Runtime.WorkerSuccessPayload, WorkerError>) async -> Void
-    let resumeQueue: @Sendable () async -> Void
-}
-
-actor PlaybackController {
+extension PlaybackQueue {
     struct ConcurrencyAdmissionThresholds {
         let startupBufferTargetMS: Int
         let concurrentGenerationTargetMS: Int
@@ -32,11 +21,6 @@ actor PlaybackController {
         let fragileOverlapWindowProgress: FragileOverlapWindowProgress?
     }
 
-    struct ActivePlayback {
-        let requestID: String
-        let task: Task<Void, Never>
-    }
-
     struct ConcurrencySnapshot: Equatable {
         let activeRequestID: String?
         let isStableForConcurrentGeneration: Bool
@@ -49,22 +33,6 @@ actor PlaybackController {
         let activeRequestID: String?
         let activeRequestTuningProfile: PlaybackTuningProfile?
         let allowsConcurrentGeneration: Bool
-    }
-
-    private let driver: AnyPlaybackController
-    private var hooks: PlaybackHooks?
-    private var activePlayback: ActivePlayback?
-    private var activePlaybackIsStableForConcurrentGeneration = false
-    private var activePlaybackStableBufferedAudioMS: Int?
-    private var activePlaybackStableBufferTargetMS: Int?
-    private var activePlaybackConcurrentGenerationTargetMS: Int?
-    private var activePlaybackFragileOverlapWindowProgress: FragileOverlapWindowProgress?
-    private var activePlaybackIsRebuffering = false
-    private var jobs = [String: LiveSpeechPlaybackState]()
-    private var queue = [String]()
-
-    init(driver: AnyPlaybackController) {
-        self.driver = driver
     }
 
     static func concurrencyAdmissionThresholds(
@@ -165,74 +133,6 @@ actor PlaybackController {
         )
     }
 
-    // MARK: - Binding
-
-    func bind(_ hooks: PlaybackHooks) {
-        self.hooks = hooks
-        Task {
-            await driver.bindEnvironmentEvents { [weak self] (event: PlaybackEnvironmentEvent) in
-                guard let self else { return }
-
-                let activeRequest = await activeRequestSummary()
-                await hooks.handleEnvironmentEvent(event, activeRequest)
-            }
-        }
-    }
-
-    // MARK: - Driver Control
-
-    func prepare(sampleRate: Double) async throws -> Bool {
-        try await driver.prepare(sampleRate: sampleRate)
-    }
-
-    func handle(_ action: PlaybackAction) async -> PlaybackState {
-        switch action {
-            case .pause:
-                let driverState = await driver.pause()
-                return resolvedPlaybackState(driverState: driverState)
-            case .resume:
-                let driverState = await driver.resume()
-                return resolvedPlaybackState(driverState: driverState)
-            case .state:
-                let driverState = await driver.state()
-                return resolvedPlaybackState(driverState: driverState)
-        }
-    }
-
-    // MARK: - Queue Ownership
-
-    func enqueue(_ request: LiveSpeechRequestState) {
-        let playbackState = LiveSpeechPlaybackState(
-            request: request,
-            execution: .make(requestID: request.id),
-        )
-        jobs[playbackState.id] = playbackState
-        queue.append(playbackState.id)
-    }
-
-    func setGenerationTask(_ task: Task<Void, Never>?, for requestID: String) {
-        jobs[requestID]?.execution.generationTask = task
-    }
-
-    func playbackState(for requestID: String) -> LiveSpeechPlaybackState? {
-        jobs[requestID]
-    }
-
-    func jobCount() -> Int {
-        jobs.count
-    }
-
-    func activeRequestSummary() -> ActiveWorkerRequestSummary? {
-        guard let requestID = activePlayback?.requestID, let playbackState = jobs[requestID] else { return nil }
-
-        return ActiveWorkerRequestSummary(
-            id: requestID,
-            kind: playbackState.request.kind,
-            voiceProfile: playbackState.request.voiceProfile,
-            requestContext: playbackState.request.requestContext,
-        )
-    }
-
     func coordinationTelemetrySnapshot() -> ConcurrencySnapshot {
         ConcurrencySnapshot(
             activeRequestID: activePlayback?.requestID,
@@ -252,205 +152,7 @@ actor PlaybackController {
         )
     }
 
-    func hasActivePlayback() -> Bool {
-        activePlayback != nil
-    }
-
-    func queuedRequestSummaries() -> [QueuedWorkerRequestSummary] {
-        let waitingQueue = queue.filter { $0 != activePlayback?.requestID }
-        return waitingQueue.enumerated().compactMap { offset, requestID -> QueuedWorkerRequestSummary? in
-            guard let playbackState = jobs[requestID] else { return nil }
-
-            return QueuedWorkerRequestSummary(
-                id: requestID,
-                kind: playbackState.request.kind,
-                voiceProfile: playbackState.request.voiceProfile,
-                requestContext: playbackState.request.requestContext,
-                queuePosition: offset + 1,
-            )
-        }
-    }
-
-    func workerStateSnapshot() async -> SpeakSwiftly.WorkerPlaybackStateSnapshot {
-        let activeRequest = activeRequestSummary()
-        let telemetry = coordinationTelemetrySnapshot()
-        let driverState = await driver.state()
-        return SpeakSwiftly.WorkerPlaybackStateSnapshot(
-            state: resolvedPlaybackState(driverState: driverState, activeRequest: activeRequest),
-            activeRequest: activeRequest,
-            isStableForConcurrentGeneration: telemetry.isStableForConcurrentGeneration,
-            isRebuffering: telemetry.isRebuffering,
-            stableBufferedAudioMS: telemetry.stableBufferedAudioMS,
-            stableBufferTargetMS: telemetry.stableBufferTargetMS,
-        )
-    }
-
-    func stateSnapshot(sequence: Int, capturedAt: Date) async -> SpeakSwiftly.PlaybackSnapshot {
-        await workerStateSnapshot().playbackSnapshot(
-            sequence: sequence,
-            capturedAt: capturedAt,
-            queuedRequests: queuedRequestSummaries(),
-        )
-    }
-
-    func clearQueued(excluding protectedRequestIDs: Set<String>) -> [LiveSpeechPlaybackState] {
-        let waitingRequestIDs = queue.filter { !protectedRequestIDs.contains($0) }
-        return waitingRequestIDs.compactMap { requestID in
-            let playbackState = jobs.removeValue(forKey: requestID)
-            queue.removeAll { $0 == requestID }
-            return playbackState
-        }
-    }
-
-    func cancel(requestID: String, cancelGenerationTask: Bool = true) async -> LiveSpeechPlaybackState? {
-        guard let playbackState = jobs[requestID] else { return nil }
-
-        if cancelGenerationTask {
-            playbackState.execution.generationTask?.cancel()
-        }
-        playbackState.execution.playbackTask?.cancel()
-        queue.removeAll { $0 == requestID }
-        jobs.removeValue(forKey: requestID)
-
-        if activePlayback?.requestID == requestID {
-            activePlayback = nil
-            await driver.stop()
-        }
-
-        return playbackState
-    }
-
-    func discard(requestID: String) -> LiveSpeechPlaybackState? {
-        queue.removeAll { $0 == requestID }
-        return jobs.removeValue(forKey: requestID)
-    }
-
-    func shutdown() async -> [LiveSpeechPlaybackState] {
-        let activeTask = activePlayback?.task
-        activePlayback = nil
-        activeTask?.cancel()
-
-        for playbackState in jobs.values {
-            playbackState.execution.generationTask?.cancel()
-            playbackState.execution.playbackTask?.cancel()
-        }
-
-        let cancelledJobs = Array(jobs.values)
-        jobs.removeAll()
-        queue.removeAll()
-        await driver.stop()
-        return cancelledJobs
-    }
-
-    // MARK: - Playback Execution
-
-    func startNextIfPossible() async {
-        guard activePlayback == nil else { return }
-        guard let requestID = queue.first, let playbackState = jobs[requestID] else { return }
-        guard let sampleRate = playbackState.execution.sampleRate else { return }
-        guard let hooks else { return }
-
-        let task = Task {
-            await self.runPlayback(for: playbackState, sampleRate: sampleRate, hooks: hooks)
-        }
-        activePlayback = ActivePlayback(requestID: requestID, task: task)
-        activePlaybackIsStableForConcurrentGeneration = false
-        activePlaybackStableBufferedAudioMS = nil
-        activePlaybackStableBufferTargetMS = nil
-        activePlaybackConcurrentGenerationTargetMS = nil
-        activePlaybackFragileOverlapWindowProgress = nil
-        activePlaybackIsRebuffering = false
-        playbackState.execution.playbackTask = task
-    }
-
-    private func resolvedPlaybackState(
-        driverState: PlaybackState,
-        activeRequest: ActiveWorkerRequestSummary? = nil,
-    ) -> PlaybackState {
-        let resolvedActiveRequest = activeRequest ?? activeRequestSummary()
-        guard resolvedActiveRequest != nil else {
-            return .idle
-        }
-
-        if driverState == .paused {
-            return .paused
-        }
-
-        return .playing
-    }
-
-    private func runPlayback(
-        for playbackState: LiveSpeechPlaybackState,
-        sampleRate: Double,
-        hooks: PlaybackHooks,
-    ) async {
-        let result: Result<SpeakSwiftly.Runtime.WorkerSuccessPayload, WorkerError>
-
-        do {
-            let playbackEngineWasPrepared = try await driver.prepare(sampleRate: sampleRate)
-            if playbackEngineWasPrepared {
-                await hooks.logEngineReady(playbackState.request, sampleRate)
-            }
-            let playbackSummary = try await driver.play(
-                sampleRate: sampleRate,
-                text: playbackState.request.normalizedText,
-                tuningProfile: playbackState.request.playbackTuningProfile,
-                stream: playbackState.execution.stream,
-            ) { event in
-                await self.recordConcurrencyEvent(event, for: playbackState.id)
-                await hooks.handleEvent(event, playbackState.request)
-            }
-            await hooks.logFinished(playbackState.request, playbackSummary, sampleRate)
-            result = .success(SpeakSwiftly.Runtime.WorkerSuccessPayload(id: playbackState.id))
-        } catch is CancellationError {
-            result = .failure(
-                WorkerError(
-                    code: .requestCancelled,
-                    message: "Request '\(playbackState.id)' was cancelled before it could complete.",
-                ),
-            )
-        } catch let workerError as WorkerError {
-            result = .failure(workerError)
-        } catch {
-            result = .failure(
-                WorkerError(
-                    code: .audioPlaybackFailed,
-                    message: "Live playback failed for request '\(playbackState.id)' due to an unexpected internal error. \(error.localizedDescription)",
-                ),
-            )
-        }
-
-        await finishPlayback(requestID: playbackState.id, result: result, hooks: hooks)
-    }
-
-    private func finishPlayback(
-        requestID: String,
-        result: Result<SpeakSwiftly.Runtime.WorkerSuccessPayload, WorkerError>,
-        hooks: PlaybackHooks,
-    ) async {
-        guard activePlayback?.requestID == requestID else { return }
-
-        activePlayback = nil
-        activePlaybackIsStableForConcurrentGeneration = false
-        activePlaybackStableBufferedAudioMS = nil
-        activePlaybackStableBufferTargetMS = nil
-        activePlaybackConcurrentGenerationTargetMS = nil
-        activePlaybackFragileOverlapWindowProgress = nil
-        activePlaybackIsRebuffering = false
-        queue.removeAll { $0 == requestID }
-
-        guard let playbackState = jobs.removeValue(forKey: requestID) else {
-            await startNextIfPossible()
-            return
-        }
-
-        playbackState.execution.generationTask = nil
-        playbackState.execution.playbackTask = nil
-        await hooks.completeJob(playbackState.request, result)
-        await hooks.resumeQueue()
-    }
-
-    private func recordConcurrencyEvent(_ event: PlaybackEvent, for requestID: String) {
+    func recordConcurrencyEvent(_ event: PlaybackEvent, for requestID: String) {
         guard activePlayback?.requestID == requestID else { return }
 
         switch event {
