+++ Sources/SpeakSwiftly/Runtime/WorkerEventModels.swift
