+++ Sources/SpeakSwiftly/Runtime/WorkerOutputModels.swift
